refactor dpm outcome review routes
